refactor: move ts types to types/ folder
