+++ src/types/database.types.ts
@@ -247,6 +247,8 @@ export interface Database {
 
 export type Tables<T extends keyof Database['public']['Tables']> = Database['public']['Tables'][T]['Row'];
 export type Enums<T extends keyof Database['public']['Enums']> = Database['public']['Enums'][T];
+export type FunctionResponse<T extends keyof Database['public']['Functions']> =
+	Database['public']['Functions'][T]['Returns'];
 
 export type DbResult<T> = T extends PromiseLike<infer U> ? U : never;
 export type DbResultOk<T> = T extends PromiseLike<{ data: infer U }> ? Exclude<U, null> : never;
