+++ src/types/global-components.d.ts
