+++ src/types/turnstile-callback.d.ts
@@ -1,3 +1,7 @@
+/* eslint-disable */
+/* prettier-ignore */
+// @ts-nocheck
+import { loadTurnstileCbName } from '../global-vars';
 import { loadTurnstileCbName } from './global-vars';
 import { TurnstileObject } from 'turnstile-types';
