+++ tsconfig.app.json
@@ -4,8 +4,9 @@
 	"exclude": ["src/**/__tests__/*"],
 	"compilerOptions": {
 		"composite": true,
-		"baseUrl": ".",
 		"tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
+
+		"baseUrl": ".",
 		"paths": {
 			"@/*": ["./src/*"]
 		},
